optimizer: make aggregate reuse constraint-free
