+++ src/include/duckdb/optimizer/aggregate_rewrite_helper.hpp
@@ -9,12 +9,14 @@
 
 #include "duckdb/planner/column_binding_map.hpp"
 #include "duckdb/planner/logical_operator.hpp"
+#include "duckdb/common/optional_idx.hpp"
 
 namespace duckdb {
 class LogicalAggregate;
 class Optimizer;
 
 struct AggregateRewriteHelper {
+	static optional_idx GetDirectReferenceIndex(const Expression &expression, LogicalOperator &input);
 	static vector<Identifier> GenerateColumnNames(const string &prefix, idx_t column_count);
 	static unique_ptr<Expression> CopyAndRebind(const Expression &expr,
 	                                            const column_binding_map_t<ColumnBinding> &replacement_map);
