+++ src/optimizer/CMakeLists.txt
@@ -12,6 +12,7 @@ add_library_unity(
   OBJECT
   aggregate_function_rewriter.cpp
   aggregate_reuse.cpp
+  materialized_aggregate_reuse.cpp
   aggregate_rewrite_helper.cpp
   build_probe_side_optimizer.cpp
   column_binding_replacer.cpp
@@ -33,7 +34,6 @@ add_library_unity(
   grouping_sets_optimizer.cpp
   in_clause_rewriter.cpp
   join_filter_pushdown_optimizer.cpp
-  key_properties.cpp
   late_materialization.cpp
   late_materialization_helper.cpp
   optimizer.cpp
