+++ src/optimizer/relation_statistics/relation_statistics_scan.cpp
@@ -241,7 +241,11 @@ idx_t RelationStatisticsHelper::InspectTableFilter(idx_t cardinality, const Tabl
 	if (expr.GetExpressionClass() == ExpressionClass::BOUND_FUNCTION) {
 		auto &function = expr.Cast<BoundFunctionExpression>();
 		if (function.Function().GetName() == "!~~" && !base_stats.CanHaveNull()) {
-			return LossyNumericCast<idx_t>(static_cast<double>(cardinality) * (1.0 - DEFAULT_SELECTIVITY));
+			if (cardinality == 0) {
+				return 0;
+			}
+			return MaxValue<idx_t>(
+			    LossyNumericCast<idx_t>(static_cast<double>(cardinality) * (1.0 - DEFAULT_SELECTIVITY)), 1);
 		}
 	}
 	if (expr.GetExpressionType() == ExpressionType::CONJUNCTION_AND) {
