+++ test/optimizer/aggregate_reuse.test
@@ -3,13 +3,13 @@
 # group: [optimizer]
 
 statement ok
-CREATE TABLE reuse_dim(id INTEGER PRIMARY KEY, label VARCHAR)
+CREATE TABLE reuse_dim(id INTEGER, label VARCHAR)
 
 statement ok
 INSERT INTO reuse_dim VALUES (10, 'x'), (20, 'b'), (30, 'x'), (40, 'd')
 
 statement ok
-CREATE TABLE reuse_owner(k INTEGER PRIMARY KEY, dim_id INTEGER)
+CREATE TABLE reuse_owner(k INTEGER, dim_id INTEGER)
 
 statement ok
 INSERT INTO reuse_owner VALUES (1, 10), (2, 20), (3, 30), (4, 40)
@@ -38,7 +38,7 @@ physical_plan	<REGEX>:(?s).*reuse_fact.*reuse_fact.*
 statement ok
 RESET disabled_optimizers
 
-# The unique key survives the dimension join, so the retained grouped payload replaces the outer scan and aggregate.
+# The grouped state replaces the outer fact scan while an upper combine preserves owner multiplicity.
 query II
 EXPLAIN SELECT o.k, d.label, sum(f.v)
 FROM reuse_owner o JOIN reuse_dim d ON o.dim_id = d.id JOIN reuse_fact f USING (k)
@@ -54,6 +54,14 @@ WHERE o.k IN (SELECT k FROM reuse_fact GROUP BY k HAVING sum(v) > 25)
 GROUP BY o.k, d.label
 ----
 physical_plan	<!REGEX>:.*Join Type: SEMI.*
+
+query II
+EXPLAIN SELECT o.k, d.label, sum(f.v)
+FROM reuse_owner o JOIN reuse_dim d ON o.dim_id = d.id JOIN reuse_fact f USING (k)
+WHERE o.k IN (SELECT k FROM reuse_fact GROUP BY k HAVING sum(v) > 25)
+GROUP BY o.k, d.label
+----
+physical_plan	<REGEX>:.*combine_aggr.*
 
 statement ok
 SET debug_verify_serializer = true
@@ -83,7 +91,7 @@ FROM reuse_owner o JOIN reuse_dim d ON o.dim_id = d.id JOIN reuse_fact f USING (
 WHERE o.k IN (SELECT k FROM reuse_fact GROUP BY k HAVING sum(v) > 25)
 GROUP BY d.label
 ----
-physical_plan	<REGEX>:(?s).*reuse_fact.*reuse_fact.*
+physical_plan	<!REGEX>:(?s).*reuse_fact.*reuse_fact.*
 
 query IT
 SELECT d.label, sum(f.v)
@@ -157,7 +165,7 @@ GROUP BY o.k, d.label
 ----
 physical_plan	<REGEX>:(?s).*reuse_fact.*reuse_fact.*
 
-# A duplicate owner would multiply the reused payload and still requires an outer aggregate.
+# Duplicate owners repeat the state and are combined by the retained outer aggregate.
 statement ok
 CREATE TABLE reuse_duplicate_owner AS SELECT * FROM reuse_owner UNION ALL SELECT * FROM reuse_owner
 
@@ -167,7 +175,15 @@ FROM reuse_duplicate_owner o JOIN reuse_fact f USING (k)
 WHERE o.k IN (SELECT k FROM reuse_fact GROUP BY k HAVING sum(v) > 25)
 GROUP BY o.k
 ----
-physical_plan	<REGEX>:(?s).*reuse_fact.*reuse_fact.*
+physical_plan	<!REGEX>:(?s).*reuse_fact.*reuse_fact.*
+
+query II
+EXPLAIN SELECT o.k, sum(f.v)
+FROM reuse_duplicate_owner o JOIN reuse_fact f USING (k)
+WHERE o.k IN (SELECT k FROM reuse_fact GROUP BY k HAVING sum(v) > 25)
+GROUP BY o.k
+----
+physical_plan	<REGEX>:.*combine_aggr.*
 
 query II rowsort
 SELECT o.k, sum(f.v)
@@ -183,7 +199,7 @@ statement ok
 SET delim_join_as_cte = true
 
 statement ok
-CREATE TABLE payload_part(k INTEGER PRIMARY KEY, size INTEGER, kind VARCHAR, label VARCHAR)
+CREATE TABLE payload_part(k INTEGER, size INTEGER, kind VARCHAR, label VARCHAR)
 
 statement ok
 CREATE TABLE payload_fact(k INTEGER, supplier_id INTEGER, cost INTEGER)
@@ -278,9 +294,9 @@ s2	a2	2	2	p2	9
 statement ok
 SET debug_verify_serializer = false
 
-# Observed uniqueness is insufficient to prove that the added relation preserves multiplicity.
+# Uniform duplication by the added relation does not change MIN.
 statement ok
-CREATE TABLE payload_unproven_part AS SELECT * FROM payload_part
+CREATE TABLE payload_unproven_part AS SELECT * FROM payload_part UNION ALL SELECT * FROM payload_part
 
 query II
 EXPLAIN WITH outer_rows AS MATERIALIZED (
@@ -294,6 +310,74 @@ EXPLAIN WITH outer_rows AS MATERIALIZED (
   GROUP BY f.k
 )
 SELECT o.k, o.cost FROM outer_rows o JOIN minimum_cost m ON o.k = m.k AND o.cost = m.cost
+----
+logical_opt	<!REGEX>:(?s).*payload_fact.*payload_fact.*
+
+query II rowsort
+WITH outer_rows AS MATERIALIZED (
+  SELECT p.k, f.cost, f.supplier_id
+  FROM payload_unproven_part p JOIN payload_fact f USING (k)
+  WHERE p.size = 15
+), minimum_cost AS (
+  SELECT f.k, min(f.cost) AS cost
+  FROM payload_fact f
+  WHERE f.k IN (SELECT k FROM outer_rows)
+  GROUP BY f.k
+)
+SELECT o.k, o.cost FROM outer_rows o JOIN minimum_cost m ON o.k = m.k AND o.cost = m.cost
+----
+1	1
+1	1
+2	9
+2	9
+2	9
+2	9
+
+# Multiple replication-invariant aggregates can be reused together.
+query II
+EXPLAIN WITH outer_rows AS MATERIALIZED (
+  SELECT p.k, f.cost, f.supplier_id
+  FROM payload_unproven_part p JOIN payload_fact f USING (k)
+  WHERE p.size = 15
+), cost_stats AS (
+  SELECT f.k, min(f.cost) AS minimum, max(f.cost) AS maximum
+  FROM payload_fact f
+  WHERE f.k IN (SELECT k FROM outer_rows)
+  GROUP BY f.k
+)
+SELECT * FROM cost_stats
+----
+logical_opt	<!REGEX>:(?s).*payload_fact.*payload_fact.*
+
+query III rowsort
+WITH outer_rows AS MATERIALIZED (
+  SELECT p.k, f.cost, f.supplier_id
+  FROM payload_unproven_part p JOIN payload_fact f USING (k)
+  WHERE p.size = 15
+), cost_stats AS (
+  SELECT f.k, min(f.cost) AS minimum, max(f.cost) AS maximum
+  FROM payload_fact f
+  WHERE f.k IN (SELECT k FROM outer_rows)
+  GROUP BY f.k
+)
+SELECT * FROM cost_stats
+----
+1	1	7
+2	9	9
+
+# Multiplicity-sensitive aggregates cannot consume the duplicated producer.
+query II
+EXPLAIN WITH outer_rows AS MATERIALIZED (
+  SELECT p.k, f.cost, f.supplier_id
+  FROM payload_unproven_part p JOIN payload_fact f USING (k)
+  WHERE p.size = 15
+), total_cost AS (
+  SELECT f.k, sum(f.cost) AS cost
+  FROM payload_fact f
+  WHERE f.k IN (SELECT k FROM outer_rows)
+  GROUP BY f.k
+)
+SELECT * FROM total_cost
 ----
 logical_opt	<REGEX>:(?s).*payload_fact.*payload_fact.*
