+++ test/optimizer/partial_aggregate_pushdown.test
@@ -1070,9 +1070,9 @@ WHERE abs(e.s - f.s) > 1e-9 * abs(f.s) + 1e-9
 ----
 0
 
-# A unique preserved key allows count to finish below an owner-outer join.
+# A high-collapse event key allows count to be partially computed below an owner-outer join.
 statement ok
-CREATE TABLE outer_count_owner(k INTEGER PRIMARY KEY, label VARCHAR)
+CREATE TABLE outer_count_owner(k INTEGER, label VARCHAR)
 
 statement ok
 INSERT INTO outer_count_owner VALUES (1, 'a'), (2, 'b'), (3, 'c'), (4, 'd')
@@ -1089,7 +1089,7 @@ ANALYZE
 
 query II
 EXPLAIN SELECT o.k, count(e.event_id)
-FROM outer_count_owner o LEFT JOIN outer_count_events e ON o.k = e.k AND e.keep
+FROM outer_count_owner o LEFT JOIN outer_count_events e ON o.k = e.k
 GROUP BY o.k
 ----
 physical_plan	<REGEX>:(?s).*COALESCE.*Hash Join.*Hash Group By.*
@@ -1099,12 +1099,12 @@ SET debug_verify_serializer = true
 
 query II rowsort
 SELECT o.k, count(e.event_id)
-FROM outer_count_owner o LEFT JOIN outer_count_events e ON o.k = e.k AND e.keep
+FROM outer_count_owner o LEFT JOIN outer_count_events e ON o.k = e.k
 GROUP BY o.k
 ----
-1	24
-2	25
-3	23
+1	30
+2	30
+3	30
 4	0
 
 statement ok
@@ -1113,7 +1113,7 @@ SET debug_verify_serializer = false
 statement ok
 CREATE TEMP TABLE outer_count_enabled AS
 SELECT o.k, count(e.event_id) c
-FROM outer_count_owner o LEFT JOIN outer_count_events e ON o.k = e.k AND e.keep
+FROM outer_count_owner o LEFT JOIN outer_count_events e ON o.k = e.k
 GROUP BY o.k
 
 statement ok
@@ -1122,7 +1122,7 @@ PRAGMA disable_optimizer
 statement ok
 CREATE TEMP TABLE outer_count_baseline AS
 SELECT o.k, count(e.event_id) c
-FROM outer_count_owner o LEFT JOIN outer_count_events e ON o.k = e.k AND e.keep
+FROM outer_count_owner o LEFT JOIN outer_count_events e ON o.k = e.k
 GROUP BY o.k
 
 statement ok
@@ -1142,7 +1142,7 @@ SET disabled_optimizers = 'partial_aggregate_pushdown'
 
 query II
 EXPLAIN SELECT o.k, count(e.event_id)
-FROM outer_count_owner o LEFT JOIN outer_count_events e ON o.k = e.k AND e.keep
+FROM outer_count_owner o LEFT JOIN outer_count_events e ON o.k = e.k
 GROUP BY o.k
 ----
 physical_plan	<!REGEX>:.*COALESCE.*
@@ -1151,10 +1151,10 @@ query I
 SELECT count(*) FROM (
 	(SELECT * FROM outer_count_enabled EXCEPT ALL
 	 SELECT o.k, count(e.event_id) FROM outer_count_owner o
-	 LEFT JOIN outer_count_events e ON o.k = e.k AND e.keep GROUP BY o.k)
+	 LEFT JOIN outer_count_events e ON o.k = e.k GROUP BY o.k)
 	UNION ALL
 	(SELECT o.k, count(e.event_id) FROM outer_count_owner o
-	 LEFT JOIN outer_count_events e ON o.k = e.k AND e.keep GROUP BY o.k
+	 LEFT JOIN outer_count_events e ON o.k = e.k GROUP BY o.k
 	 EXCEPT ALL SELECT * FROM outer_count_enabled)
 ) differences
 ----
@@ -1163,9 +1163,12 @@ SELECT count(*) FROM (
 statement ok
 RESET disabled_optimizers
 
-# An unproven owner key, extra grouping column, and other aggregates retain the established plan.
+# Owner uniqueness is not required: the upper SUM preserves duplicate owner multiplicity.
 statement ok
-CREATE TABLE outer_count_unproven AS SELECT * FROM outer_count_owner
+CREATE TABLE outer_count_unproven AS SELECT * FROM outer_count_owner UNION ALL SELECT * FROM outer_count_owner
+
+statement ok
+ANALYZE outer_count_unproven
 
 statement ok
 CREATE TABLE outer_count_nullable(k INTEGER UNIQUE, label VARCHAR)
@@ -1182,15 +1185,72 @@ ANALYZE outer_count_sparse_events
 
 query II
 EXPLAIN SELECT o.k, count(e.event_id)
-FROM outer_count_unproven o LEFT JOIN outer_count_events e ON o.k = e.k AND e.keep
+FROM outer_count_unproven o LEFT JOIN outer_count_events e ON o.k = e.k
 GROUP BY o.k
 ----
-physical_plan	<!REGEX>:.*COALESCE.*
+physical_plan	<REGEX>:.*COALESCE.*
+
+query II rowsort
+SELECT o.k, count(e.event_id)
+FROM outer_count_unproven o LEFT JOIN outer_count_events e ON o.k = e.k
+GROUP BY o.k
+----
+1	60
+2	60
+3	60
+4	0
+
+# COUNT(*) contributes one for an owner row without a match.
+query II
+EXPLAIN SELECT o.k, count(*)
+FROM outer_count_owner o LEFT JOIN outer_count_events e ON o.k = e.k
+GROUP BY o.k
+----
+physical_plan	<REGEX>:.*COALESCE.*
+
+query II rowsort
+SELECT o.k, count(*)
+FROM outer_count_owner o LEFT JOIN outer_count_events e ON o.k = e.k
+GROUP BY o.k
+----
+1	34
+2	33
+3	33
+4	1
+
+query II
+EXPLAIN SELECT o.k, count(*)
+FROM outer_count_events e RIGHT JOIN outer_count_unproven o ON e.k = o.k
+GROUP BY o.k
+----
+physical_plan	<REGEX>:.*COALESCE.*
+
+query II rowsort
+SELECT o.k, count(*)
+FROM outer_count_events e RIGHT JOIN outer_count_unproven o ON e.k = o.k
+GROUP BY o.k
+----
+1	68
+2	66
+3	66
+4	2
 
 # Preaggregation is rejected when the owner join would discard most event rows.
 query II
 EXPLAIN SELECT o.k, count(e.event_id)
 FROM outer_count_owner o LEFT JOIN outer_count_sparse_events e ON o.k = e.k
+GROUP BY o.k
+----
+physical_plan	<!REGEX>:.*COALESCE.*
+
+# The estimated group count must account for a large non-owner key domain.
+query II
+EXPLAIN SELECT o.k, count(e.event_id)
+FROM outer_count_owner o
+LEFT JOIN (
+  SELECT CASE WHEN i < 7000 THEN 1 ELSE i END AS k, i AS event_id
+  FROM range(10000) t(i)
+) e ON o.k = e.k
 GROUP BY o.k
 ----
 physical_plan	<!REGEX>:.*COALESCE.*
@@ -1202,17 +1262,27 @@ GROUP BY o.k
 ----
 physical_plan	<!REGEX>:.*COALESCE.*
 
+# NULL owner keys retain their unmatched count semantics when reliable key statistics are available.
+query II
+EXPLAIN SELECT o.k, count(e.event_id)
+FROM outer_count_nullable o LEFT JOIN outer_count_events e ON o.k = e.k
+GROUP BY o.k
+----
+physical_plan	<REGEX>:.*COALESCE.*
+
+query II rowsort
+SELECT o.k, count(e.event_id)
+FROM outer_count_nullable o LEFT JOIN outer_count_events e ON o.k = e.k
+GROUP BY o.k
+----
+1	30
+2	30
+NULL	0
+
 query II
 EXPLAIN SELECT o.k, o.label, count(e.event_id)
 FROM outer_count_owner o LEFT JOIN outer_count_events e ON o.k = e.k AND e.keep
 GROUP BY o.k, o.label
-----
-physical_plan	<!REGEX>:.*COALESCE.*
-
-query II
-EXPLAIN SELECT o.k, count(*)
-FROM outer_count_owner o LEFT JOIN outer_count_events e ON o.k = e.k AND e.keep
-GROUP BY o.k
 ----
 physical_plan	<!REGEX>:.*COALESCE.*
