+++ test/sql/optimizer/test_negated_filter_cardinality.test
@@ -12,6 +12,12 @@ statement ok
 CREATE TABLE nullable_strings AS SELECT CASE WHEN i = 0 THEN NULL ELSE 'ordinary' END AS value FROM range(1000) t(i);
 
 statement ok
+CREATE TABLE one_string(value VARCHAR NOT NULL);
+
+statement ok
+INSERT INTO one_string VALUES ('ordinary');
+
+statement ok
 PRAGMA disable_verification;
 
 statement ok
@@ -28,6 +34,12 @@ query II
 EXPLAIN (FORMAT JSON) SELECT * FROM strings WHERE value NOT LIKE '%special%requests%';
 ----
 logical_opt	<REGEX>:.*"name":\s*"SEQ_SCAN".*"Estimated Cardinality":\s*"800".*
+
+# A non-empty input must not be rounded down to zero rows.
+query II
+EXPLAIN (FORMAT JSON) SELECT * FROM one_string WHERE value NOT LIKE '%special%requests%';
+----
+logical_opt	<REGEX>:.*"name":\s*"SEQ_SCAN".*"Estimated Cardinality":\s*"1".*
 
 # A nullable input cannot use the complement because NULL satisfies neither LIKE nor NOT LIKE.
 query II
